+++ shell/src/context_listener.rs
@@ -38,7 +38,13 @@ impl ContextListener {
     ///
     /// This actor spawns a new thread in which it listens for incoming events from the `protocol_runner`.
     /// Events are received from IPC channel provided by [`event_server`](IpcEvtServer).
-    pub fn actor(sys: &impl ActorRefFactory, persistent_storage: &PersistentStorage, mut event_server: IpcEvtServer, log: Logger) -> Result<ContextListenerRef, CreateError> {
+    pub fn actor(
+        sys: &impl ActorRefFactory,
+        persistent_storage: &PersistentStorage,
+        mut event_server: IpcEvtServer,
+        log: Logger,
+        store_context_action: bool
+    ) -> Result<ContextListenerRef, CreateError> {
         let context_storage = persistent_storage.context_storage();
         let listener_run = Arc::new(AtomicBool::new(true));
         let block_applier_thread = {
@@ -55,6 +61,7 @@ impl ContextListener {
                         &mut context_action_storage,
                         &mut context,
                         &log,
+                        store_context_action,
                     ) {
                         Ok(()) => debug!(log, "Context listener finished"),
                         Err(err) => {
@@ -106,12 +113,31 @@ impl Actor for ContextListener {
     }
 }
 
+fn store_action(storage: &mut ContextActionStorage, should_store: bool, action: ContextAction) -> Result<(), Error> {
+    if !should_store { return Ok(()); }
+    match &action {
+        ContextAction::Set { block_hash: Some(block_hash), .. }
+        | ContextAction::Copy { block_hash: Some(block_hash), .. }
+        | ContextAction::Delete { block_hash: Some(block_hash), .. }
+        | ContextAction::RemoveRecursively { block_hash: Some(block_hash), .. }
+        | ContextAction::Mem { block_hash: Some(block_hash), .. }
+        | ContextAction::DirMem { block_hash: Some(block_hash), .. }
+        | ContextAction::Get { block_hash: Some(block_hash), .. }
+        | ContextAction::Fold { block_hash: Some(block_hash), .. } => {
+            storage.put_action(&block_hash.clone(), action)?;
+            Ok(())
+        }
+        _ => Ok(()),
+    }
+}
+
 fn listen_protocol_events(
     apply_block_run: &AtomicBool,
     event_server: &mut IpcEvtServer,
     context_action_storage: &mut ContextActionStorage,
     context: &mut Box<dyn ContextApi>,
     log: &Logger,
+    store_context_actions: bool,
 ) -> Result<(), Error> {
     debug!(log, "Waiting for connection from protocol runner");
     let mut rx = event_server.accept()?;
@@ -139,45 +165,24 @@ fn listen_protocol_events(
                 event_count += 1;
 
                 match &msg {
-                    ContextAction::Set { block_hash: Some(block_hash), key, value, context_hash, ignored, .. } => {
-                        if !ignored {
-                            context_diff.set(context_hash, key, value)?;
-                        }
-                        context_action_storage.put_action(&block_hash.clone(), msg)?;
-                    }
-                    ContextAction::Copy { block_hash: Some(block_hash), to_key: key, from_key, context_hash, ignored, .. } => {
-                        if !ignored {
-                            context.copy_to_diff(context_hash, from_key, key, &mut context_diff)?;
-                        }
-                        context_action_storage.put_action(&block_hash.clone(), msg)?;
-                    }
-                    | ContextAction::Delete { block_hash: Some(block_hash), key, context_hash, ignored, .. } => {
-                        if !ignored {
-                            context.delete_to_diff(context_hash, key, &mut context_diff)?;
-                        }
-                        context_action_storage.put_action(&block_hash.clone(), msg)?;
-                    }
-                    | ContextAction::RemoveRecursively { block_hash: Some(block_hash), key, context_hash, ignored, .. } => {
-                        if !ignored {
-                            context.remove_recursively_to_diff(context_hash, key, &mut context_diff)?;
-                        }
-                        context_action_storage.put_action(&block_hash.clone(), msg)?;
-                    }
-                    ContextAction::Commit { parent_context_hash, new_context_hash, block_hash: Some(block_hash), .. } => {
-                        context.commit(block_hash, parent_context_hash, new_context_hash, &context_diff)?;
-                    }
+                    ContextAction::Set { key, value, context_hash, ignored, .. } =>
+                        if !ignored { context_diff.set(context_hash, key, value)?; }
+                    ContextAction::Copy { to_key: key, from_key, context_hash, ignored, .. } =>
+                        if !ignored { context.copy_to_diff(context_hash, from_key, key, &mut context_diff)?; }
+                    ContextAction::Delete { key, context_hash, ignored, .. } =>
+                        if !ignored { context.delete_to_diff(context_hash, key, &mut context_diff)?; }
+                    ContextAction::RemoveRecursively { key, context_hash, ignored, .. } =>
+                        if !ignored { context.remove_recursively_to_diff(context_hash, key, &mut context_diff)?; }
+                    ContextAction::Commit { parent_context_hash, new_context_hash, block_hash: Some(block_hash), .. } =>
+                        context.commit(block_hash, parent_context_hash, new_context_hash, &context_diff)?,
                     ContextAction::Checkout { context_hash, .. } => {
-                        context_diff = context.checkout(context_hash)?;
                         event_count = 0;
-                    }
-                    ContextAction::Mem { block_hash: Some(block_hash), .. }
-                    | ContextAction::DirMem { block_hash: Some(block_hash), .. }
-                    | ContextAction::Get { block_hash: Some(block_hash), .. }
-                    | ContextAction::Fold { block_hash: Some(block_hash), .. } => {
-                        context_action_storage.put_action(&block_hash.clone(), msg)?;
+                        context_diff = context.checkout(context_hash)?;
                     }
                     _ => (),
                 };
+
+                store_action(context_action_storage, store_context_actions, msg)?;
             }
             Err(err) => {
                 warn!(log, "Failed to receive event from protocol runner"; "reason" => format!("{:?}", err));