+++ tezos/interop/build.rs
@@ -14,7 +14,7 @@ use serde_json;
 use sha2::{Digest, Sha256};
 
 const GIT_REPO_URL: &str = "https://gitlab.com/simplestaking/tezos.git";
-const GIT_COMMIT_HASH: &str = "d8747c1dcefe51aa50bd1052bf2333dfe77fc703";
+const GIT_COMMIT_HASH: &str = "51a4dacdc8ec139b9ed61f7dd97deec1d7d40a9b";
 const GIT_RELEASE_DISTRIBUTIONS_FILE: &str = "lib_tezos/libtezos-ffi-distribution-summary.json";
 const GIT_REPO_DIR: &str = "lib_tezos/src";