+++ CHANGELOG.md
@@ -31,6 +31,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Nothing.
 
+## [0.0.1] - 2020-05-18
+
+### Added
+
+- Flags --store-p2p-messages=BOOL and --store-context-actions=BOOL. If these flags are set to false,
+the node will persist less data to disk, which increases runtime speed. 
+ 
 ## [0.0.1] - 2020-03-31
 
 ### Added